requirements.txt
Flask==3.1.1
gunicorn==23.0.0

requests
cos-python-sdk-v5
tencentcloud-sdk-python